refactor: move `ReportForm`
