+++ packages/frontend/src/components/Form/ReportForm.tsx
@@ -1,24 +1,23 @@
 import { FormEvent, useState, useMemo, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { CenterModal } from '../../Modals/CenterModal'
-import FeedbackButton from '../../Buttons/FeedbackButton/FeedbackButton'
-import { FeedbackForm } from '../FeedbackForm/FeedbackForm'
-import { SelectField } from '../SelectField/SelectField'
-import { Line } from '../../Miscellaneous/Line/Line'
+import { CenterModal } from '../Modals/CenterModal'
+import FeedbackButton from '../Buttons/FeedbackButton/FeedbackButton'
+import { FeedbackForm } from './FeedbackForm/FeedbackForm'
+import { SelectField } from './SelectField/SelectField'
+import { Line } from '../Miscellaneous/Line/Line'
 import { useLines, useStations } from 'src/api/queries'
 import { Report, Station } from 'src/utils/types'
 import { getClosestStations } from 'src/hooks/getClosestStations'
 import { useLocation } from 'src/contexts/LocationContext'
 import { useStationSearch } from 'src/hooks/useStationSearch'
 import { validateReport, ValidationError } from 'src/utils/reportValidation'
-import searchIcon from '../../../../public/icons/search.svg'
-import StationButton from '../../Buttons/StationButton'
-import { Link } from 'react-router-dom'
+import searchIcon from '../../../public/icons/search.svg'
+import StationButton from '../Buttons/StationButton'
 import { sendAnalyticsEvent } from 'src/hooks/useAnalytics'
 import { useSubmitReport } from 'src/api/queries'
-import { TextAreaWithPrivacy, TextAreaWithPrivacyRef } from '../TextAreaWithPrivacy/TextAreaWithPrivacy'
-import { SubmitButton } from '../../common/SubmitButton/SubmitButton'
+import { TextAreaWithPrivacy, TextAreaWithPrivacyRef } from './TextAreaWithPrivacy/TextAreaWithPrivacy'
+import { SubmitButton } from '../common/SubmitButton/SubmitButton'
 
 interface ReportFormProps {
     onReportFormSubmit: (reportedData: Report) => void
